Added first test to Test_Pathfinders, testing first case for GoTo class. Added a return for GoTo find_path, just to check tests.
