--- Tests/Test-Pathfinders.py
@@ -1,18 +1,0 @@
-import unittest
-from Pathfinders import GoTo
-
-
-class TestPathfinders(unittest.TestCase):
-
-    def test_goto_no_obstacles(self):
-        pass
-
-    def test_goto_obstacle(self):
-        pass
-
-    def test_goto_no_valid_path(self):
-        pass
-
-
-if __name__ == "__main__":
-    unittest.main()
