+++ Tests/ToolsForTesting.py
@@ -1,0 +1,3 @@
+def calculate_manhattan_distance(position1, position2):
+    result = abs(position1[0] - position2[0]) + abs(position1[1] - position2[1])
+    return result
